Agregando pruebas y aplicando correcciones
